+++ mockserver-core/src/test/java/org/mockserver/serialization/serializers/response/HttpResponseDTOSerializerTest.java
@@ -101,7 +101,11 @@ public class HttpResponseDTOSerializerTest {
                 "  \"cookies\" : {" + NEW_LINE +
                 "    \"cookie_name\" : \"cookie_value\"" + NEW_LINE +
                 "  }," + NEW_LINE +
-                "  \"body\" : \"{ \\\"key\\\": \\\"some_value\\\" }\"" + NEW_LINE +
+                "  \"body\" : {" + NEW_LINE +
+                "    \"type\" : \"JSON\"," + NEW_LINE +
+                "    \"json\" : \"{ \\\"key\\\": \\\"some_value\\\" }\"," + NEW_LINE +
+                "    \"rawBytes\" : \"eyAia2V5IjogInNvbWVfdmFsdWUiIH0=\"" + NEW_LINE +
+                "  }" + NEW_LINE +
                 "}"));
     }