+++ changelog.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - reduced object creation (and therefore GCs) for log especially during request matching
 - print logs timestamp with milliseconds  
 - reduced expiry of certification to one year to avoid errors from modern systems that don't like long lived certificates (such as Chrome or VMWare)
+- defaulted charset for XML and JSON to UTF8 as per rfc3470 and rfc8259
 
 ### Fixed
 - fixed but with environment vairable configuration for long, integer and integer list values